requirements.txt
# Core dependencies
torch>=2.0.0
torchvision>=0.15.0
transformers>=4.37.0
accelerate>=0.25.0

# Data processing
numpy>=1.24.0
pandas>=2.0.0
rasterio>=1.3.0
openpyxl>=3.1.0
scikit-learn>=1.3.0
scipy>=1.11.0
joblib>=1.3.0

# Training utilities
deepspeed>=0.12.0
peft>=0.7.0  # For LoRA
trl>=0.7.0   # For DPO/RLHF
bitsandbytes>=0.41.0  # For quantization

# Logging and monitoring
wandb>=0.16.0
tensorboard>=2.15.0
tqdm>=4.66.0

# Configuration
pyyaml>=6.0
hydra-core>=1.3.0

# Visualization (optional)
matplotlib>=3.8.0
seaborn>=0.13.0

# Development
pytest>=7.4.0
black>=23.0.0
isort>=5.12.0
